Fix: moved anything about TdiRefZone under TdiCompile's lock, i.e. add_compile_info
Conflicts:
	tdishr/TdiCompile.c
	tdishr/TdiIntrinsic.c
