+++ tdishr/TdiCompile.c
@@ -28,8 +28,12 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
         Ken Klare, LANL CTR-7   (c)1989,1990
 */
 
-extern unsigned short OpcCompile;
 #define DEF_FREEXD
+extern unsigned short OpcCompile, OpcEvaluate;
+#define MAXLINE 120
+#define MAXFRAC 40
+#define MINMAX(min, test, max) ((min) >= (test) ? (min) : (test) < (max) ? (test) : (max))
+#define MAXMESS 1800
 
 #include <stdlib.h>
 #include <string.h>
@@ -37,10 +41,11 @@ extern unsigned short OpcCompile;
 #include "tdirefzone.h"
 #include "tdirefstandard.h"
 #include <libroutines.h>
+#include <treeshr_messages.h>
 #include <tdishr_messages.h>
 #include "tdithreadsafe.h"
 #include <mdsshr.h>
-#include <STATICdef.h>
+#include <strroutines.h>
 
 extern void LockMdsShrMutex(pthread_mutex_t *, int *);
 extern void UnlockMdsShrMutex(pthread_mutex_t *);
@@ -66,9 +71,94 @@ extern void TdiYyReset();
         IMMEDIATE (`) must never call COMPILE. NEED to prevent this.
 */
 
-int Tdi1Compile(int opcode __attribute__ ((unused)), int narg, struct descriptor *list[],
-		struct descriptor_xd *out_ptr)
-{
+static pthread_mutex_t lock = PTHREAD_MUTEX_INITIALIZER;
+static void cleanup_compile(ThreadStatic * TdiThreadStatic_p){
+  if (TdiRefZone.a_begin) {
+    free(TdiRefZone.a_begin);
+    TdiRefZone.a_begin=NULL;
+  }
+  pthread_mutex_unlock(&lock);
+}
+
+static inline void add_compile_info(int status,ThreadStatic * TdiThreadStatic_p) {
+  if(!(status==TdiSYNTAX
+    || status==TdiEXTRANEOUS
+    || status==TdiUNBALANCE
+    || status==TreeNOT_OPEN
+    || status==TreeNNF
+    || status==TdiBOMB)) return;
+  static const DESCRIPTOR(compile_err, "%TDI Error compiling region marked by ^\n");
+  struct descriptor_d *message = &TdiThreadStatic_p->TdiIntrinsic_message;
+  if (!TdiRefZone.a_begin || message->length >= MAXMESS)
+    return;
+  // b------x----c----e
+  // '-l_ok-'-xc-'-ce-'
+  char *b = TdiRefZone.a_begin;
+  char *e = TdiRefZone.a_end;
+  char *c = MINMAX(b, TdiRefZone.a_cur, e);
+  char *x = MINMAX(b, b + TdiRefZone.l_ok, c);
+  int xc = MINMAX(0, c-x, MAXLINE);
+  int ce = MINMAX(0, e-c, MAXLINE);
+  int bx = MINMAX(0, x-b, MAXLINE);
+  if (xc+ce > MAXLINE)
+    ce = MINMAX(0, ce, MAXFRAC);
+  if (bx+xc+ce > MAXLINE)
+    bx = MINMAX(0, bx, MAXFRAC);
+  int len = bx+xc+2;
+  struct descriptor marker = { len, DTYPE_T, CLASS_S, memset(malloc(len+1),' ',len) };
+  struct descriptor region = { bx+xc+ce, DTYPE_T, CLASS_S, x-bx };
+  marker.pointer[bx]='^';
+  marker.pointer[0] = marker.pointer[len-1]='\n';
+  if (xc>0)
+    marker.pointer[bx+xc]='^';
+  else if (bx>0)
+    marker.pointer[bx]='^';
+  StrAppend(message,(struct descriptor *)&compile_err);
+  StrAppend(message,(struct descriptor *)&region);
+  StrAppend(message,(struct descriptor *)&marker);
+  free(marker.pointer);
+}
+
+static inline int tdi_compile(ThreadStatic * TdiThreadStatic_p,struct descriptor * text_ptr, int narg, struct descriptor *list[], struct descriptor_xd *out_ptr) {
+  int status;
+  TdiThreadStatic_p->compiler_recursing = 1;
+  pthread_mutex_lock(&lock);
+  pthread_cleanup_push((void*)cleanup_compile,(void*)TdiThreadStatic_p)
+  if (TdiRefZone.l_zone) status = SsSUCCESS;
+  else status = LibCreateVmZone(&TdiRefZone.l_zone);
+ /****************************************
+  In case we bomb out, probably not needed.
+  ****************************************/
+  TdiRefZone.l_status = TdiBOMB;
+  if (TdiRefZone.a_begin) free(TdiRefZone.a_begin);
+  TdiRefZone.a_begin = TdiRefZone.a_cur = memcpy(malloc(text_ptr->length), text_ptr->pointer, text_ptr->length);
+  TdiRefZone.a_end = TdiRefZone.a_cur + text_ptr->length;
+  TdiRefZone.l_ok = 0;
+  TdiRefZone.l_narg = narg - 1;
+  TdiRefZone.l_iarg = 0;
+  TdiRefZone.a_list = &list[0];
+  TdiYyReset();
+  if (IS_NOT_OK(TdiYacc()) && IS_OK(TdiRefZone.l_status))
+    status = TdiSYNTAX;
+  else
+    status = TdiRefZone.l_status;
+ /************************
+  Move from temporary zone.
+  ************************/
+  if STATUS_OK {
+    if (TdiRefZone.a_result == 0)
+      MdsFree1Dx(out_ptr, NULL);
+    else
+      status = MdsCopyDxXd((struct descriptor *)TdiRefZone.a_result, out_ptr);
+  }
+  LibResetVmZone(&TdiRefZone.l_zone);
+  add_compile_info(status,TdiThreadStatic_p);
+  pthread_cleanup_pop(1);
+  TdiThreadStatic_p->compiler_recursing = 0;
+  return status;
+}
+
+EXPORT int Tdi1Compile(int opcode, int narg, struct descriptor *list[], struct descriptor_xd *out_ptr){
   int status;
   struct descriptor *text_ptr;
   GET_TDITHREADSTATIC_P;
@@ -78,53 +168,16 @@ int Tdi1Compile(int opcode __attribute__ ((unused)), int narg, struct descriptor
   }
   EMPTYXD(tmp);
   FREEXD_ON_EXIT(&tmp);
-  status = TdiEvaluate(list[0], &tmp MDS_END_ARG);
+  status = Tdi1Evaluate(opcode, 1, list, &tmp);// using Tdi1Evaluate over TdiEvaluate saves 3 stack levels
   text_ptr = tmp.pointer;
   if (STATUS_OK && text_ptr->dtype != DTYPE_T)
     status = TdiINVDTYDSC;
   if STATUS_OK {
-    if (text_ptr->length > 0) {
-      TdiThreadStatic_p->compiler_recursing = 1;
-      static pthread_mutex_t lock = PTHREAD_MUTEX_INITIALIZER; // TODO: try to avoid
-      pthread_mutex_lock(&lock);
-      pthread_cleanup_push((void*)pthread_mutex_unlock,&lock);
-      if (!TdiRefZone.l_zone)
-	status = LibCreateVmZone(&TdiRefZone.l_zone);
-      /****************************************
-      In case we bomb out, probably not needed.
-      ****************************************/
-      TdiRefZone.l_status = TdiBOMB;
-      if (TdiRefZone.a_begin) free(TdiRefZone.a_begin);
-      TdiRefZone.a_begin = TdiRefZone.a_cur = memcpy(malloc(text_ptr->length), text_ptr->pointer, text_ptr->length);
-      TdiRefZone.a_end = TdiRefZone.a_cur + text_ptr->length;
-      TdiRefZone.l_ok = 0;
-      TdiRefZone.l_narg = narg - 1;
-      TdiRefZone.l_iarg = 0;
-      TdiRefZone.a_list = &list[0];
-      if STATUS_OK {
-	TdiYyReset();
-	if (IS_NOT_OK(TdiYacc()) && IS_OK(TdiRefZone.l_status))
-	  status = TdiSYNTAX;
-	else
-	  status = TdiRefZone.l_status;
-      }
-      /************************
-      Move from temporary zone.
-      ************************/
-      if STATUS_OK {
-	if (TdiRefZone.a_result == 0)
-	  MdsFree1Dx(out_ptr, NULL);
-	else
-	  status = MdsCopyDxXd((struct descriptor *)TdiRefZone.a_result, out_ptr);
-      }
-      LibResetVmZone(&TdiRefZone.l_zone);
-      pthread_cleanup_pop(1);
-      TdiThreadStatic_p->compiler_recursing = 0;
-    }
+    if (text_ptr->length > 0)
+      status = tdi_compile(TdiThreadStatic_p,text_ptr,narg,list,out_ptr);
   }
   FREEXD_NOW(&tmp);
   if STATUS_NOT_OK MdsFree1Dx(out_ptr, NULL);
-
   return status;
 }
